refactor(config): split config.go and rename fields to camelCase
- Extract postgresPortFromEnv, loadEnvFile, findEnvFile into loader.go
- Extract validateDatabaseConfig into validation.go
- Rename AppConfig fields from SCREAMING_SNAKE to camelCase (PostgresHost etc.)
- Skip .env loading in production (APP_ENV=production)
- Harden JWT_SECRET: fatal in production, warn with safe default in dev
- Update database/db.go to use renamed AppConfig fields
